計算結果が浮動小数点になるバグ修正

+++ novel_app/lib/screens/novel_generate/novel_generate_screen.dart
@@ -173,7 +173,7 @@ class _NovelGenerateScreenState extends State<NovelGenerateScreen> {
                                       (selectedTime *
                                               selectedUnit *
                                               450 *
-                                              selectedReadingSpeed)
+                                              selectedReadingSpeed.toInt())
                                           .toString(),
                                   genre: selectedGenre,
                                   style: selectedStyle,
